+++ tests/NCronJob.Tests/JobOptionBuilderTests.cs
@@ -107,6 +107,96 @@ public class JobOptionBuilderTests
     }
 
     [Fact]
+    public void ShouldCreateMultipleNamedAndScheduledJobsWithAnd()
+    {
+        var builder = new JobOptionBuilder();
+        builder
+            .WithName("name1")
+            .WithCronExpression(Cron.AtEveryMinute)
+            .And
+            .WithName("name2")
+            .WithCronExpression(Cron.AtEveryMinute);
+
+        var options = builder.GetJobOptions();
+
+        options.Count.ShouldBe(2);
+        options[0].Name.ShouldBe("name1");
+        options[0].CronExpression.ShouldBe(Cron.AtEveryMinute);
+        options[1].Name.ShouldBe("name2");
+        options[1].CronExpression.ShouldBe(Cron.AtEveryMinute);
+    }
+
+    [Fact]
+    public void ShouldCreateMultipleNamedAndScheduledJobsWithoutAnd()
+    {
+        var builder = new JobOptionBuilder();
+        builder
+            .WithName("name1")
+            .WithCronExpression(Cron.AtEveryMinute);
+
+        builder
+            .WithName("name2")
+            .WithCronExpression(Cron.AtEveryMinute);
+
+        var options = builder.GetJobOptions();
+
+        options.Count.ShouldBe(2);
+        options[0].Name.ShouldBe("name1");
+        options[0].CronExpression.ShouldBe(Cron.AtEveryMinute);
+        options[1].Name.ShouldBe("name2");
+        options[1].CronExpression.ShouldBe(Cron.AtEveryMinute);
+    }
+
+    [Fact]
+    public void ShouldCreateMultipleNamedScheduledAndParameterizedJobsWithAnd()
+    {
+        var builder = new JobOptionBuilder();
+        builder
+            .WithName("name1")
+            .WithCronExpression(Cron.AtEveryMinute)
+            .WithParameter("foo")
+            .And
+            .WithName("name2")
+            .WithCronExpression(Cron.AtEveryMinute)
+            .WithParameter("bar");
+
+        var options = builder.GetJobOptions();
+
+        options.Count.ShouldBe(2);
+        options[0].Name.ShouldBe("name1");
+        options[0].CronExpression.ShouldBe(Cron.AtEveryMinute);
+        options[0].Parameter.ShouldBe("foo");
+        options[1].Name.ShouldBe("name2");
+        options[1].CronExpression.ShouldBe(Cron.AtEveryMinute);
+        options[1].Parameter.ShouldBe("bar");
+    }
+
+    [Fact]
+    public void ShouldCreateMultipleNamedScheduledAndParameterizedJobsWithoutAnd()
+    {
+        var builder = new JobOptionBuilder();
+        builder
+            .WithName("name1")
+            .WithCronExpression(Cron.AtEveryMinute)
+            .WithParameter("foo");
+
+        builder
+            .WithName("name2")
+            .WithCronExpression(Cron.AtEveryMinute)
+            .WithParameter("bar");
+
+        var options = builder.GetJobOptions();
+
+        options.Count.ShouldBe(2);
+        options[0].Name.ShouldBe("name1");
+        options[0].CronExpression.ShouldBe(Cron.AtEveryMinute);
+        options[0].Parameter.ShouldBe("foo");
+        options[1].Name.ShouldBe("name2");
+        options[1].CronExpression.ShouldBe(Cron.AtEveryMinute);
+        options[1].Parameter.ShouldBe("bar");
+    }
+
+    [Fact]
     public void ShouldCreateMultipleNamedAndParameterizedJobsWithAnd()
     {
         var builder = new JobOptionBuilder();